+++ tests/testthat/test_fitting.R
@@ -10,8 +10,8 @@ test_that("Test data stimulation step", {
   # create a realistic looking spectrum
   set.seed(1)
   
-  mrs_data <- sim_res$mrs_data %>% lb(6) %>% "/"(10) %>% phase(130) %>%
-              shift(0.1) %>% add_noise(0.001, fd = FALSE)
+  mrs_data <- sim_res$mrs_data |> lb(6) |> scale_mrs_amp(1 / 10) |>
+              phase(130) |> shift(0.1) |> add_noise(0.001, fd = FALSE)
   
   expect_equal_to_reference(mrs_data, "abfit_sim_mrs_data.rds",
                             tolerance = 1e-6)
@@ -37,7 +37,7 @@ test_that("Test ABfit without iterative optimisation", {
   opts     <- abfit_opts(maxiters_pre = 0, maxiters = 0)
   
   # undo the large phase error to avoid FWHM measurement warnings
-  fit_res  <- fit_mrs(mrs_data %>% phase(-130), sim_res$basis, method = "abfit",
+  fit_res  <- fit_mrs(mrs_data |> phase(-130), sim_res$basis, method = "abfit",
                       opts = opts, time = FALSE, progress = "none")
   
   expect_equal_to_reference(fit_res, "abfit_res_no_optim.rds", tolerance = 1e-5)
@@ -49,8 +49,8 @@ test_that("Test ABfit coarse-fitting steps only", {
   
   # don't run the fine fitting step
   opts <- abfit_opts(maxiters = 0)
-  fit_res  <- fit_mrs(mrs_data, sim_res$basis, method = "abfit", opts = opts,
-                      time = FALSE, progress = "none")
+  fit_res <- fit_mrs(mrs_data, sim_res$basis, method = "abfit", opts = opts,
+                     time = FALSE, progress = "none")
   
   # this test is a bit more sensitive to differences between platforms
   expect_equal_to_reference(fit_res, "abfit_res_coarse.rds", tolerance = 2e-1)
@@ -66,7 +66,7 @@ test_that("Test ABfit fine-fitting only", {
   opts <- abfit_opts(maxiters_pre = 0)
   
   # undo the large phase error to avoid FWHM measurement warnings
-  fit_res  <- fit_mrs(mrs_data %>% phase(-120), sim_res$basis, method = "abfit",
+  fit_res  <- fit_mrs(mrs_data |> phase(-120), sim_res$basis, method = "abfit",
                       opts = opts, time = FALSE, progress = "none")
   
   expect_equal_to_reference(fit_res, "abfit_res_fine.rds", tolerance = 1e-6)